ShardCounter int --> long
git-svn-id: https://everest.teltel.com:8443/repos/misc/9x9dev/trunk/nncloudtv@5496 a0a5f996-2bb4-401c-8778-395c3c01700f

+++ src/main/java/com/nncloudtv/dao/ShardedCounter.java
@@ -104,8 +104,8 @@ public class ShardedCounter {
     return counterStored;
   }
 
-  public int getCount() {
-    int sum = 0;
+  public long getCount() {
+    long sum = 0;
     PersistenceManager pm = PMF.getContent().getPersistenceManager();
 
     try {
